STM32 Tiny: Rename files for better conformance to naming conventions
