+++ nuttx/configs/stm32_tiny/src/stm32_leds.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * configs/stm32_tiny/src/up_leds.c
+ * configs/stm32_tiny/src/stm32_leds.c
  *
  *   Copyright (C) 2009, 2011, 2013, 2015 Gregory Nutt. All rights reserved.
  *   Author: Laurent Latil <laurent@latil.nom.fr>
