+++ nuttx/configs/stm32_tiny/src/stm32_spi.c
@@ -1,5 +1,5 @@
 /************************************************************************************
- * configs/stm32_tiny/src/up_spi.c
+ * configs/stm32_tiny/src/stm32_spi.c
  *
  *   Copyright (C) 2009, 2011-2013 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
@@ -55,7 +55,7 @@
 #if defined(CONFIG_STM32_SPI1) || defined(CONFIG_STM32_SPI2)
 
 /************************************************************************************
- * Definitions
+ * Pre-processor Definitions
  ************************************************************************************/
 
 /* Enables debug output from this file (needs CONFIG_DEBUG too) */
