+++ nuttx/configs/stm32_tiny/src/stm32_watchdog.c
@@ -1,5 +1,5 @@
 /************************************************************************************
- * configs/stm32_tiny/src/up_pwm.c
+ * configs/stm_tiny/src/stm32_watchdog.c
  *
  *   Copyright (C) 2012-2013 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
@@ -42,24 +42,62 @@
 #include <errno.h>
 #include <debug.h>
 
-#include <nuttx/pwm.h>
+#include <nuttx/watchdog.h>
 #include <arch/board/board.h>
 
-#include "chip.h"
-#include "up_arch.h"
-#include "stm32_pwm.h"
-#include "stm32_tiny-internal.h"
+#include "stm32_wdg.h"
+
+#ifdef CONFIG_WATCHDOG
 
 /************************************************************************************
- * Definitions
+ * Pre-processor Definitions
  ************************************************************************************/
 /* Configuration *******************************************************************/
-/* PWM
- *
- * The STM32 Tiny board provides a LED on GPIO line B5.
- */
+/* Wathdog hardware should be enabled */
 
-#ifdef CONFIG_PWM
+#if !defined(CONFIG_STM32_WWDG) && !defined(CONFIG_STM32_IWDG)
+#  warning "One of CONFIG_STM32_WWDG or CONFIG_STM32_IWDG must be defined"
+#endif
+
+/* Select the path to the registered watchdog timer device */
+
+#ifndef CONFIG_STM32_WDG_DEVPATH
+#  ifdef CONFIG_EXAMPLES_WATCHDOG_DEVPATH
+#    define CONFIG_STM32_WDG_DEVPATH CONFIG_EXAMPLES_WATCHDOG_DEVPATH
+#  else
+#    define CONFIG_STM32_WDG_DEVPATH "/dev/watchdog0"
+#  endif
+#endif
+
+/* Use the un-calibrated LSI frequency if we have nothing better */
+
+#if defined(CONFIG_STM32_IWDG) && !defined(CONFIG_STM32_LSIFREQ)
+#  define CONFIG_STM32_LSIFREQ STM32_LSI_FREQUENCY
+#endif
+
+/* Debug ***************************************************************************/
+/* Non-standard debug that may be enabled just for testing the watchdog timer */
+
+#ifndef CONFIG_DEBUG
+#  undef CONFIG_DEBUG_WATCHDOG
+#endif
+
+#ifdef CONFIG_DEBUG_WATCHDOG
+#  define wdgdbg                 dbg
+#  define wdglldbg               lldbg
+#  ifdef CONFIG_DEBUG_VERBOSE
+#    define wdgvdbg              vdbg
+#    define wdgllvdbg            llvdbg
+#  else
+#    define wdgvdbg(x...)
+#    define wdgllvdbg(x...)
+#  endif
+#else
+#  define wdgdbg(x...)
+#  define wdglldbg(x...)
+#  define wdgvdbg(x...)
+#  define wdgllvdbg(x...)
+#endif
 
 /************************************************************************************
  * Private Functions
@@ -69,49 +107,29 @@
  * Public Functions
  ************************************************************************************/
 
-/************************************************************************************
- * Name: pwm_devinit
+/****************************************************************************
+ * Name: up_wdginitialize()
  *
  * Description:
- *   All STM32 architectures must provide the following interface to work with
- *   examples/pwm.
+ *   Perform architecuture-specific initialization of the Watchdog hardware.
+ *   This interface must be provided by all configurations using
+ *   apps/examples/watchdog
  *
- ************************************************************************************/
+ ****************************************************************************/
 
-int pwm_devinit(void)
+int up_wdginitialize(void)
 {
-  static bool initialized = false;
-  struct pwm_lowerhalf_s *pwm;
-  int ret;
+  /* Initialize tha register the watchdog timer device */
 
-  /* Have we already initialized? */
-
-  if (!initialized)
-    {
-      /* Call stm32_pwminitialize() to get an instance of the PWM interface */
-
-      pwm = stm32_pwminitialize(STM32TINY_PWMTIMER);
-      if (!pwm)
-        {
-          adbg("Failed to get the STM32 PWM lower half\n");
-          return -ENODEV;
-        }
-
-      /* Register the PWM driver at "/dev/pwm0" */
-
-      ret = pwm_register("/dev/pwm0", pwm);
-      if (ret < 0)
-        {
-          adbg("pwm_register failed: %d\n", ret);
-          return ret;
-        }
-
-      /* Now we are initialized */
-
-      initialized = true;
-    }
-
+#if defined(CONFIG_STM32_WWDG)
+  stm32_wwdginitialize(CONFIG_STM32_WDG_DEVPATH);
   return OK;
+#elif defined(CONFIG_STM32_IWDG)
+  stm32_iwdginitialize(CONFIG_STM32_WDG_DEVPATH, CONFIG_STM32_LSIFREQ);
+  return OK;
+#else
+  return -ENODEV;
+#endif
 }
 
-#endif /* CONFIG_PWM */
+#endif /* CONFIG_WATCHDOG */
